Allow custom resolution caps on the Weston output
As there is no way to reasonably implement "maximum usable res"
with weston, Raspberry Pi configs, or any other approach, this
implements some logic for Weston startup to ensure that the given
resolution or lower is used.

This is useful to e.g. ensure that when you have a variety of
displays that range from 720p to 4k and you want to ensure that
the WPE block will use the native resolution on all but the 4K
screens where it should use the 1080p resolution (as the rpi4 is
too slow to reasonably play video content on those), you can
actually do that, without requiring manually specifying the
res for each device.